added selectable store sales and lot of improvements

+++ NettbutikkAPI/NettbutikkAPI/Order.cs
@@ -4,11 +4,11 @@ public class Order
 {
     public int UserId { get; private set; }
     public int OrderId { get; private set; }
-    public int TotalPrice { get; private set; }
+    public double TotalPrice { get; private set; }
     public List<Product> OrderItems { get; private set; }
     public int Status { get; private set; }
     
-    public Order(int userId, int orderId, int totalPrice, List<Product> orderItems, int status)
+    public Order(int userId, int orderId, double totalPrice, List<Product> orderItems, int status)
     {
         UserId = userId;
         OrderId = orderId;
